+++ src/state.rs
@@ -1,16 +1,12 @@
 use std::collections::BTreeMap;
-use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
-use std::time::SystemTime;
 
-use chrono::prelude::*;
-use log::{debug, info};
 use serde_derive::{Deserialize, Serialize};
 
 use crate::error::Error;
 use crate::settings::Settings;
-use crate::task::{Task, TaskResult, TaskStatus};
+use crate::task::{Task, TaskStatus};
 
 pub type SharedState = Arc<Mutex<State>>;
 
@@ -53,7 +49,9 @@ pub struct State {
     pub tasks: BTreeMap<usize, Task>,
     /// All groups
     pub groups: BTreeMap<String, GroupStatus>,
-    config_path: Option<PathBuf>,
+    /// Used to store an configuration path that has been explicitely specified.
+    /// Without this, the default config path will be used instead.
+    pub config_path: Option<PathBuf>,
 }
 
 impl State {
@@ -126,8 +124,7 @@ impl State {
             }
         }
 
-        self.save()?;
-        self.save_settings()
+        Ok(())
     }
 
     /// Set the group status (running/paused) for all groups including the default queue.
@@ -136,16 +133,6 @@ impl State {
         for key in keys {
             self.groups.insert(key, status.clone());
         }
-    }
-
-    /// Get all ids of task with a specific state inside a specific group.
-    pub fn task_ids_in_group_with_stati(&self, group: &str, stati: Vec<TaskStatus>) -> Vec<usize> {
-        self.tasks
-            .iter()
-            .filter(|(_, task)| stati.contains(&task.status))
-            .filter(|(_, task)| task.group.eq(group))
-            .map(|(id, _)| *id)
-            .collect()
     }
 
     /// Get all ids of task inside a specific group.
@@ -214,7 +201,7 @@ impl State {
         // Filter all task id's that match the provided statuses.
         for task_id in task_ids.iter() {
             // Check whether the task exists and save all non-existing task ids.
-            match self.tasks.get(&task_id) {
+            match self.tasks.get(task_id) {
                 None => {
                     mismatching.push(*task_id);
                     continue;
@@ -231,232 +218,5 @@ impl State {
         }
 
         (matching, mismatching)
-    }
-
-    /// Check if a task can be deleted. \
-    /// We have to check all dependant tasks, that haven't finished yet.
-    /// This is necessary to prevent deletion of tasks which are specified as a dependency.
-    ///
-    /// `to_delete` A list of task ids, which should also be deleted.
-    ///             This allows to remove dependency tasks as well as their dependants.
-    pub fn is_task_removable(&self, task_id: &usize, to_delete: &[usize]) -> bool {
-        // Get all task ids of any dependant tasks.
-        let dependants: Vec<usize> = self
-            .tasks
-            .iter()
-            .filter(|(_, task)| {
-                task.dependencies.contains(&task_id) && matches!(task.status, TaskStatus::Done(_))
-            })
-            .map(|(_, task)| task.id)
-            .collect();
-
-        if dependants.is_empty() {
-            return true;
-        }
-
-        // Check if the dependants are supposed to be deleted as well.
-        let should_delete_dependants = dependants.iter().all(|task_id| to_delete.contains(task_id));
-        if !should_delete_dependants {
-            return false;
-        }
-
-        // Lastly, do a recursive check if there are any dependants on our dependants
-        dependants
-            .iter()
-            .all(|task_id| self.is_task_removable(task_id, to_delete))
-    }
-
-    /// A small helper for handling task failures. \
-    /// Users can specify whether they want to pause the task's group or the
-    /// whole daemon on a failed tasks. This function wraps that logic and decides if anything should be
-    /// paused depending on the current settings.
-    ///
-    /// `group` should be the name of the failed task.
-    pub fn handle_task_failure(&mut self, group: String) {
-        if self.settings.daemon.pause_group_on_failure {
-            self.groups.insert(group, GroupStatus::Paused);
-        } else if self.settings.daemon.pause_all_on_failure {
-            self.set_status_for_all_groups(GroupStatus::Paused);
-        }
-    }
-
-    /// Do a full reset of the state.
-    /// This doesn't reset any processes!
-    pub fn reset(&mut self) -> Result<(), Error> {
-        self.backup()?;
-        self.tasks = BTreeMap::new();
-        self.set_status_for_all_groups(GroupStatus::Running);
-
-        self.save()
-    }
-
-    /// A small convenience wrapper for saving the settings to a file.
-    pub fn save_settings(&self) -> Result<(), Error> {
-        self.settings.save(&self.config_path)
-    }
-
-    /// Convenience wrapper around save_to_file.
-    pub fn save(&self) -> Result<(), Error> {
-        self.save_to_file(false)
-    }
-
-    /// Save the current current state in a file with a timestamp.
-    /// At the same time remove old state logs from the log directory.
-    /// This function is called, when large changes to the state are applied, e.g. clean/reset.
-    pub fn backup(&self) -> Result<(), Error> {
-        self.save_to_file(true)?;
-        self.rotate()
-            .map_err(|err| Error::Generic(format!("Failed to rotate old log files:\n{}", err)))?;
-        Ok(())
-    }
-
-    /// Save the current state to disk. \
-    /// We do this to restore in case of a crash. \
-    /// If log == true, the file will be saved with a time stamp.
-    ///
-    /// In comparison to the daemon -> client communication, the state is saved
-    /// as JSON for better readability and debug purposes.
-    fn save_to_file(&self, log: bool) -> Result<(), Error> {
-        let serialized = serde_json::to_string(&self);
-        if let Err(error) = serialized {
-            return Err(Error::StateSave(format!(
-                "Failed to serialize state:\n\n{}",
-                error
-            )));
-        }
-
-        let serialized = serialized.unwrap();
-        let path = self.settings.shared.pueue_directory();
-        let (temp, real) = if log {
-            let path = path.join("log");
-            let now: DateTime<Utc> = Utc::now();
-            let time = now.format("%Y-%m-%d_%H-%M-%S");
-            (
-                path.join(format!("{}_state.json.partial", time)),
-                path.join(format!("{}_state.json", time)),
-            )
-        } else {
-            (path.join("state.json.partial"), path.join("state.json"))
-        };
-
-        // Write to temporary log file first, to prevent loss due to crashes.
-        fs::write(&temp, serialized)
-            .map_err(|err| Error::StateSave(format!("Failed to write file:\n\n{}", err)))?;
-
-        // Overwrite the original with the temp file, if everything went fine.
-        fs::rename(&temp, &real).map_err(|err| {
-            Error::StateSave(format!("Failed to overwrite old log file:\n\n{}", err))
-        })?;
-
-        if log {
-            debug!("State backup created at: {:?}", real);
-        } else {
-            debug!("State saved at: {:?}", real);
-        }
-
-        Ok(())
-    }
-
-    /// Restore the last state from a previous session. \
-    /// The state is stored as json in the `pueue_directory`.
-    ///
-    /// If the state cannot be deserialized, an empty default state will be used instead. \
-    /// All groups with queued tasks will be automatically paused to prevent unwanted execution.
-    pub fn restore(&mut self) -> Result<(), Error> {
-        let path = Path::new(&self.settings.shared.pueue_directory()).join("state.json");
-
-        // Ignore if the file doesn't exist. It doesn't have to.
-        if !path.exists() {
-            info!(
-                "Couldn't find state from previous session at location: {:?}",
-                path
-            );
-            return Ok(());
-        }
-        info!("Start restoring state");
-
-        // Try to load the file.
-        let data = fs::read_to_string(&path)
-            .map_err(|err| Error::StateSave(format!("Failed to read file:\n\n{}", err)))?;
-
-        // Try to deserialize the state file.
-        let mut state: State = serde_json::from_str(&data)
-            .map_err(|err| Error::StateDeserialization(err.to_string()))?;
-
-        // Copy group statuses from the previous state.
-        for (group, _) in state.settings.daemon.groups {
-            if let Some(status) = state.groups.get(&group) {
-                self.groups.insert(group.clone(), status.clone());
-            }
-        }
-
-        // Restore all tasks.
-        // While restoring the tasks, check for any invalid/broken stati.
-        for (task_id, task) in state.tasks.iter_mut() {
-            // Handle ungraceful shutdowns while executing tasks.
-            if task.status == TaskStatus::Running || task.status == TaskStatus::Paused {
-                info!(
-                    "Setting task {} with previous status {:?} to new status {:?}",
-                    task.id,
-                    task.status,
-                    TaskResult::Killed
-                );
-                task.status = TaskStatus::Done(TaskResult::Killed);
-            }
-
-            // Handle crash during editing of the task command.
-            if task.status == TaskStatus::Locked {
-                task.status = TaskStatus::Stashed { enqueue_at: None };
-            }
-
-            // Go trough all tasks and set all groups that are no longer
-            // listed in the configuration file to the default.
-            if !self.settings.daemon.groups.contains_key(&task.group) {
-                task.set_default_group();
-            }
-
-            // If there are any queued tasks, pause the group.
-            // This should prevent any unwanted execution of tasks due to a system crash.
-            if task.status == TaskStatus::Queued {
-                info!(
-                    "Pausing group {} to prevent unwanted execution of previous tasks",
-                    &task.group
-                );
-                self.groups.insert(task.group.clone(), GroupStatus::Paused);
-            }
-
-            self.tasks.insert(*task_id, task.clone());
-        }
-
-        Ok(())
-    }
-
-    /// Remove old logs that aren't needed any longer.
-    fn rotate(&self) -> Result<(), Error> {
-        let path = self.settings.shared.pueue_directory().join("log");
-
-        // Get all log files in the directory with their respective system time.
-        let mut entries: BTreeMap<SystemTime, PathBuf> = BTreeMap::new();
-        let mut directory_list = fs::read_dir(path)?;
-        while let Some(Ok(entry)) = directory_list.next() {
-            let path = entry.path();
-
-            let metadata = entry.metadata()?;
-            let time = metadata.modified()?;
-            entries.insert(time, path);
-        }
-
-        // Remove all files above the threshold.
-        // Old files are removed first (implictly by the BTree order).
-        let mut number_entries = entries.len();
-        let mut iter = entries.iter();
-        while number_entries > 10 {
-            if let Some((_, path)) = iter.next() {
-                fs::remove_file(path)?;
-                number_entries -= 1;
-            }
-        }
-
-        Ok(())
     }
 }